Clean existing files and use the same numbering in all folders
Previously, the DiskFileSource would use different file numbers in each
folders, e.g. for folders A and B, the numbering would be A/0000.bin,
then B/0001.bin, then A/0002.bin. In a situation where the folder contents
were swapped (for example, the folders were on two separate NVMe drives
that were mounted in a random order every time the machine booted), the
existing files would be left untouched (since they would not correspond
to the numbering) and the application would likely run out of storage
space.

This commit makes the numbering the same in all folders, e.g. for folders
A and B, it would be A/0000.bin, then B/0000.bin, then A/0001.bin ; this
ensures that even if the folders were swapped, the data would still be
recognizer.

In addition, any files found in the folders that match the naming scheme
of the library, but are not part of the numbering for the current run,
will be deleted. This makes it safe to use this new version of the library
without clearing existing stores.

Upgrade 1.2.1 -> 1.3.0